fix: make some plugin time filed default null (#6813)
* fix: make some plugin time filed default null

* fix: some plugins e2e test

* fix: jira epic e2e test

+++ backend/plugins/gitlab/tasks/issue_convertor.go
@@ -78,7 +78,7 @@ func ConvertIssues(taskCtx plugin.SubTaskContext) errors.Error {
 				Description:             issue.Body,
 				Priority:                issue.Priority,
 				OriginalType:            issue.Type,
-				LeadTimeMinutes:         int64(issue.LeadTimeMinutes),
+				LeadTimeMinutes:         issue.LeadTimeMinutes,
 				Url:                     issue.Url,
 				CreatedDate:             &issue.GitlabCreatedAt,
 				UpdatedDate:             &issue.GitlabUpdatedAt,
